optimize migration

+++ src/Migrate.php
@@ -10,6 +10,7 @@
 namespace FastD\Migration;
 
 
+use FastD\QueryBuilder\MySqlBuilder;
 use PDO;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Command\HelpCommand;
@@ -32,12 +33,17 @@ class Migrate extends Command
      */
     protected $configFile;
 
+    /**
+     * @var PDO
+     */
+    protected $connection;
+
     public function configure()
     {
         $this
             ->setName('migrate')
             ->setDescription('Migration database to php')
-            ->addArgument('behavior', InputArgument::OPTIONAL, 'Migration behavior <comment>[info|dump|run|cache-clear]</comment>', 'help')
+            ->addArgument('behavior', InputArgument::OPTIONAL, 'Migration behavior <comment>[info|create|dump|run|cache-clear]</comment>', 'help')
             ->addArgument('table', InputArgument::OPTIONAL, 'Migration table name', null)
             ->addOption('conf', 'c', InputOption::VALUE_OPTIONAL, 'Config file', './migrate.yml')
             ->addOption('path', 'p', InputOption::VALUE_OPTIONAL, 'Dump or run into tables path', './seed')
@@ -52,19 +58,23 @@ class Migrate extends Command
      */
     protected function createConnection(array $config = null)
     {
-        if (null === $config) {
-            if (!file_exists($this->configFile)) {
-                throw new \RuntimeException('cannot such config file '.$this->configFile);
+        if (null === $this->connection) {
+            if (null === $config) {
+                if (!file_exists($this->configFile)) {
+                    throw new \RuntimeException('cannot such config file '.$this->configFile);
+                }
+
+                $config = Yaml::parse(file_get_contents($this->configFile));
             }
 
-            $config = Yaml::parse(file_get_contents($this->configFile));
+            $this->connection = new PDO(
+                sprintf('mysql:host=%s;dbname=%s', $config['host'], $config['dbname']),
+                $config['user'],
+                $config['pass']
+            );
         }
 
-        return new PDO(
-            sprintf('mysql:host=%s;dbname=%s', $config['host'], $config['dbname']),
-            $config['user'],
-            $config['pass']
-        );
+        return $this->connection;
     }
 
     /**
@@ -106,8 +116,9 @@ class Migrate extends Command
 
     public function verbosity(OutputInterface $output, Table $table)
     {
-        if ($output->getVerbosity() >= OutputInterface::VERBOSITY_DEBUG) {
-            $output->writeln('verbosity');
+        if ($output->getVerbosity() == OutputInterface::VERBOSITY_DEBUG) {
+            $builder = new TableBuilder($this->createConnection());
+            $output->writeln(sprintf("SQL: \n%s", $builder->update($table)->getTableInfo()));
         }
     }
 
@@ -130,6 +141,10 @@ class Migrate extends Command
         }
 
         switch ($input->getArgument('behavior')) {
+            case 'create':
+                $output->writeln($config);
+                $this->create($input, $output);
+                break;
             case 'info':
                 $output->writeln($config);
                 $this->info($input, $output);
@@ -164,8 +179,7 @@ class Migrate extends Command
      */
     protected function renderTableInfo(InputInterface $input, OutputInterface $output, Table $table)
     {
-        if ($input->hasParameterOption(['--info', '-i'])) {
-            $output->writeln(sprintf('Table: <comment>%s</comment>', $table->getTableName()));
+        if ($input->hasParameterOption(['--info', '-i']) || !empty($input->getArgument('table'))) {
             $t = new SymfonyTable($output);
             $t->setHeaders(array('Field', 'Type', 'Nullable', 'Key', 'Default', 'Extra'));
             foreach ($table->getColumns() as $column) {
@@ -205,13 +219,37 @@ class Migrate extends Command
      * @param InputInterface $input
      * @param OutputInterface $output
      */
+    public function create(InputInterface $input, OutputInterface $output)
+    {
+        $table = new Table($input->getArgument('table'));
+        $table
+            ->addColumn('id', 'int', null, false, 0, '')
+            ->addColumn('created_at', 'datetime', null, false, 'CURRENT_TIMESTAMP', '')
+            ->addColumn('updated_at', 'datetime', null, false, 'CURRENT_TIMESTAMP', '')
+        ;
+        $content = $this->dumpPhpFile($table);
+        $path = $input->getOption('path');
+        $this->targetDirectory($path);
+        $name = $this->classRename($table->getTableName());
+        $file = $path . '/' . $name . '.php';
+        file_put_contents($file, $content);
+        $output->writeln(sprintf('  <info>✔</info> Table <info>"%s"</info> <comment>dumping</comment> <info>done.</info>', $table->getTableName()));
+    }
+
+    /**
+     * @param InputInterface $input
+     * @param OutputInterface $output
+     */
     public function info(InputInterface $input, OutputInterface $output)
     {
         $builder = new TableBuilder($this->createConnection());
 
-        $tables = $builder->extract($input->getArgument('table'));
+        $tableName = $input->getArgument('table');
+
+        $tables = $builder->extract($tableName);
 
         foreach ($tables as $table) {
+            $output->writeln(sprintf('Table: <comment>%s</comment>', $table->getTableName()));
             $this->renderTableInfo($input, $output, $table);
         }
     }
@@ -225,10 +263,15 @@ class Migrate extends Command
     public function cacheClear(InputInterface $input, OutputInterface $output)
     {
         $cachePath = __DIR__ . '/.cache/tables';
-        $output->writeln(sprintf('Cache path: <comment>%s</comment>', $cachePath));
-        foreach (glob($cachePath . '/*') as $file) {
-            unlink($file);
-            $output->writeln(sprintf('  <info>✔</info> Table <info>"%s"</info> <comment>cache is clean</comment> <info>done.</info>', pathinfo($file, PATHINFO_FILENAME)));
+        $output->writeln(sprintf('Cache path: <info>%s</info>', $cachePath));
+        $files = glob($cachePath . '/*');
+        if (!empty($files)) {
+            foreach ($files as $file) {
+                unlink($file);
+                $output->writeln(sprintf('  <info>✔</info> Table <info>"%s"</info> cache is clean <info>done.</info>', pathinfo($file, PATHINFO_FILENAME)));
+            }
+        } else {
+            $output->writeln(sprintf('  <comment>!!</comment> Empty cache.'));
         }
     }
 
@@ -249,17 +292,32 @@ class Migrate extends Command
             if ($migration instanceof MigrationAbstract) {
                 $table = $migration->setUp();
                 try {
-                    $this->renderTableInfo($input, $output, $table);
-                    // not change
                     if ('' === ($sql = $builder->update($table)->getTableInfo())) {
                         $output->writeln(sprintf('  <comment>!!</comment> Table <info>"%s"</info> <comment>no change.</comment>', $table->getTableName()));
                     } else {
                         $builder->update($table)->execute();
                         $output->writeln(sprintf('  <info>✔</info> Table <info>"%s"</info> <comment>migrating</comment> <info>done.</info>', $table->getTableName()));
                     }
+                    if (($dataPath = $input->getOption('data'))) {
+                        $tableName = $table->getTableName();
+                        $dataFile = $dataPath . '/' . $tableName . '.yml';
+                        $rowsCount = 0;
+                        if (file_exists($dataFile)) {
+                            $dataset = Yaml::parse(file_get_contents($dataFile));
+                            foreach ($dataset as $row) {
+                                $sql = (new MySqlBuilder($tableName))->insert($row);
+                                if ($this->connection->exec($sql) > 0) {
+                                    $rowsCount++;
+                                }
+                            }
+                            $output->writeln(sprintf('  <info>✔</info> Table <info>"%s"</info> insert data: <info>%s</info>', $tableName, $rowsCount));
+                        }
+                    }
+                    $this->renderTableInfo($input, $output, $table);
                 } catch (\PDOException $e) {
                     $output->writeln(sprintf("<fg=red>✗</> %s \n  File: %s\n  Line: %s\n", $e->getMessage(), $e->getFile(), $e->getLine()));
                 }
+                $this->verbosity($output, $table);
             } else {
                 $output->writeln(sprintf('  <comment>!!</comment> Warning: Migrate class "<comment>%s</comment>" is not implement "<comment>%s</comment>"', $className, MigrationAbstract::class));
             }
@@ -296,7 +354,6 @@ class Migrate extends Command
                 $output->writeln(sprintf('  <comment>!!</comment> Dump table "<comment>%s</comment>" is <comment>not change</comment>', $table->getTableName()));
             }
             $this->renderTableInfo($input, $output, $table);
-            $this->verbosity($output, $table);
         }
     }
 
@@ -328,7 +385,6 @@ class Migrate extends Command
 <?php
 
 use \FastD\Migration\MigrationAbstract;
-use \FastD\Migration\Column;
 use \FastD\Migration\Table;
 
 
@@ -344,14 +400,6 @@ class {$name} extends MigrationAbstract
         {$codeString}
 
         return \$table;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function dataSet()
-    {
-        
     }
 }
 MIGRATION;
